+++ flowtxt-infrastructure/src/main/java/ca/flowtxt/infrastructure/config/InfrastructureConfig.java
@@ -1,5 +1,7 @@
 package ca.flowtxt.infrastructure.config;
 
+import ca.flowtxt.infrastructure.config.properties.RateLimitProperties;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
@@ -8,6 +10,7 @@ import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSeriali
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 
 @Configuration
+@EnableConfigurationProperties(RateLimitProperties.class)
 public class InfrastructureConfig {
 
     @Bean